feat: SQLite 연동, Steam API, AI 점수 개선, 프론트엔드 UI 완성
- SQLite 로컬 개발 지원 (JSONB→JSON, nullslast 제거, GIN 인덱스 조건부)
- NullCacheService: Redis 없이도 동작하는 캐시 서비스
- Steam 무료 API 연동 (게임 할인 데이터 자동 수집)
- AI 점수 알고리즘 개선: 가격 이력 부족 시 대안 스코어링 (할인율/키워드/카테고리 중앙값)
- 프론트엔드: 카테고리 필터, 쇼핑몰 필터, 정렬, 검색 실제 API 연결
- Decimal→float 직렬화, CORS 포트 확장

+++ backend/app/schemas/deal.py
@@ -1,7 +1,6 @@
 """Deal Pydantic schemas for request/response validation."""
 
 from datetime import datetime
-from decimal import Decimal
 from typing import List, Optional
 from uuid import UUID
 
@@ -35,10 +34,10 @@ class DealResponse(BaseModel):
 
     id: UUID
     title: str
-    deal_price: Decimal
-    original_price: Optional[Decimal] = None
-    discount_percentage: Optional[Decimal] = None
-    ai_score: Optional[Decimal] = None
+    deal_price: float
+    original_price: Optional[float] = None
+    discount_percentage: Optional[float] = None
+    ai_score: Optional[float] = None
     ai_reasoning: Optional[str] = None
     deal_type: str
     deal_url: str
